refactor: move `test-utils` folder to root of 'src' folder
