+++ packages/store-zustand/src/actions/deleteSelectedCards/__tests__/deleteSelectedCards.spec.ts
@@ -1,7 +1,6 @@
 import { AppState, TestCardId } from '../../../types';
 import { useStore } from '../../../store/zustandStore';
-
-import { defaultInitialState, generateCardsSlice } from '../../__test-utils__';
+import { defaultInitialState, generateCardsSlice } from '../../../test-utils';
 
 import { deleteSelectedCards } from '../deleteSelectedCards';
