+++ packages/store-zustand/src/actions/onDragSelection/__tests__/onDragSelection.spec.ts
@@ -1,7 +1,6 @@
 import { AppState, TestCardId, TestLayerId } from '../../../types';
 import { useStore } from '../../../store/zustandStore';
-
-import { defaultInitialState, generateCardsSlice } from '../../__test-utils__';
+import { defaultInitialState, generateCardsSlice } from '../../../test-utils';
 
 import { onDragSelection } from '../onDragSelection';
