+++ packages/store-zustand/src/actions/setCardStateToHovered/__tests__/setCardStateToHovered.spec.ts
@@ -1,7 +1,6 @@
 import { AppState, TestCardId } from '../../../types';
 import { useStore } from '../../../store/zustandStore';
-
-import { defaultInitialState, generateCardsSlice } from '../../__test-utils__';
+import { defaultInitialState, generateCardsSlice } from '../../../test-utils';
 
 import { setCardStateToHovered } from '../setCardStateToHovered';
