+++ packages/store-zustand/src/actions/setLayerStateToHovered/__tests__/setLayerStateToHovered.spec.ts
@@ -1,7 +1,6 @@
 import { AppState, LayerState, TestCardId, TestLayerId } from '../../../types';
 import { useStore } from '../../../store/zustandStore';
-
-import { defaultInitialState, generateCardsSlice } from '../../__test-utils__';
+import { defaultInitialState, generateCardsSlice } from '../../../test-utils';
 
 import { setLayerStateToHovered } from '../setLayerStateToHovered';
