+++ packages/store-zustand/src/actions/setToolToText/__tests__/setToolToText.spec.ts
@@ -1,7 +1,6 @@
 import { AppState, StoryboardTool } from '../../../types';
 import { useStore } from '../../../store/zustandStore';
-
-import { defaultInitialState } from '../../__test-utils__';
+import { defaultInitialState } from '../../../test-utils';
 
 import { setToolToText } from '../setToolToText';
