+++ packages/store-zustand/src/test-utils/generateCardsSlice/__tests__/generateCardsSlice.spec.ts
@@ -1,4 +1,4 @@
-import { TestCardId, TestLayerId } from '../../../../types';
+import { TestCardId, TestLayerId } from '../../../types';
 
 import { generateCardsSlice } from '../generateCardsSlice';
